requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.4
djangorestframework==3.15.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.0
rest-framework-simplejwt==0.0.2
rpds-py==0.18.0
simplejwt==2.0.1
six==1.16.0
sqlparse==0.5.0
typing==3.7.4.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
